ルーティングマイグレして表示できるようにする
