+++ svelte.config.js
@@ -1,33 +1,32 @@
-import adapter from '@sveltejs/adapter-auto';
+import path from 'path';
+import adapterAuto from '@sveltejs/adapter-auto';
+import adapterNode from '@sveltejs/adapter-node';
 import sveltePreprocess from 'svelte-preprocess';
 const preprocess = sveltePreprocess({
   typescript: true,
   // ...
 });
-import path from 'path';
+let adapter = null;
+
+if (adapterAuto().name !== '@sveltejs/adapter-auto') {
+  // vercel/netlify/cloudflare pages
+  adapter = adapterAuto();
+}
+else {
+  // auto で hit しなかったら強制的に node にする
+  adapter = adapterNode();
+}
+
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
   kit: {
-    adapter: adapter(),
+    adapter,
 
-    // Override http methods in the Todo forms
-    methodOverride: {
-      allowed: ['PATCH', 'DELETE']
+    alias: {
+      'svelte-modal-manager': path.resolve('./src/lib'),
+      $components: path.resolve('./src/components'),
     },
-
-    vite: {
-      mode: process.env.NODE_ENV,
-      define: {
-        'process.env': process.env,
-      },
-      resolve: {
-        alias: {
-          'svelte-modal-manager': path.resolve('./src/lib'),
-          $components: path.resolve('./src/components')
-        }
-      }
-    }
   },
   preprocess,
 };
